+++ packages/sdk/src/remittance/modules/BasicBRC29.ts
@@ -344,7 +344,7 @@ export class Brc29RemittanceModule implements RemittanceModule<
       }
 
       const keyID = `${settlement.customInstructions.derivationPrefix} ${settlement.customInstructions.derivationSuffix}`
-      const keyRequest = { protocolID, keyID, counterparty: sender }
+      const keyRequest = { protocolID, keyID, counterparty: sender, forSelf: true }
       const { publicKey } = validateWalletResult(
         'getPublicKey',
         await wallet.getPublicKey(keyRequest, origin),